add VariantToSubmittedVariantEntityProcessor test and delete unused processor
